fixup! Convert new/edit/back links into buttons
Leave the URL of the back button in /accounts/settings/edit as :back
instead of user_path(@user).

The :back symbol[1] is a "true" back that tries to take the user to
the last page they were on (which might not be the user profile page,
since the profile can be edited from the menu while on any page).

[1] https://apidock.com/rails/ActionView/Helpers/UrlHelper/link_to