fixup! Convert new/edit/back links into buttons
Remove the new 'Edit' button (/accounts/settings/edit) added to the
user profile of the current user during the conversion of links to
buttons.

Admins previously had an 'Edit' link (/user/:id/edit) on all user
profile pages. The new 'Edit' button replaces this link on the admin's
own profile.

It's not obvious that this is the right thing to do, even though
/accounts/settings/edit is mostly[1] a superset of /user/:id/edit,
because admins will no longer be able to visit /user/:id/edit for
their own account through their profile page, and the new 'Edit'
button is inconsistent with the 'Edit' link in their row on /users.

Revert to a straightforward conversion for now.

[1] /accounts/settings/edit doesn't have a brownie points field